clean: remove unused ParserHelper, move other helpers out of __tests__ directories so jest doesn't see them as tests
