--- src/parser/core/__tests__/ParserHelper.ts
@@ -1,4 +1,0 @@
-jest.mock('parser/core/Parser')
-import Parser from 'parser/core/Parser'
-
-export const MockedParser = Parser as jest.Mock<Parser>
